backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
PyJWT
cryptography
email-validator
pymysql
pypdf
python-multipart
httpx
